+++ tests/fixtures/noqa.py
@@ -581,6 +581,13 @@ def nested_annotation_return_func() -> Union[str, Union[int, float]]: # noqa: WP
     """Union[str, Union[int, float]]"""
 
 
+def optional_union_annotation_func(arg: Optional[Union[str, int]]):  # noqa: WPS704
+    """Optional[Union[str, int]]"""
+
+
+def optional_union_annotation_return_func() -> Optional[Union[str, int]]:  # noqa: WPS704
+    """Optional[Union[str, int]]"""
+
 from json import loads  # noqa: WPS347
 from some_module import a  # noqa: WPS347
 from text import from_file  # noqa: WPS347