+++ CHANGELOG.md
@@ -24,6 +24,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use redundant subscripts (e.g., `[0:7]` or `[3:None]`)
 - Allow `super()` as a valid overused expression
 - `WPS350` enforce using augmented assign pattern
+- Forbids to use `Optional[Union` in function annotations 
 
 ### Bugfixes